Infrastructure as Code/user_registration/requirements.txt
Flask
mysql-connector
flask_restful